requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.40.0
numpy>=1.24.0
matplotlib>=3.10.5
Pillow>=9.5.0
gradio>=3.50.0